Time: 7 ms (70.54%) | Memory: 25.8 MB (98.40%) - LeetSync

+++ 554-brick-wall/brick-wall.cpp
@@ -4,7 +4,7 @@ public:
 
         int totalWidth = wall.size();
         int totalHeight = wall.size();
-        long long sum;
+        uint sum;
         int maxi = 0;
         unordered_map<int, int> map;
